refactor: merge qr service to url service
