+++ internal/url/url_service_test.go
@@ -204,3 +204,35 @@ func TestFetchUserURLHistory(t *testing.T) {
 		})
 	}
 }
+
+func TestGenerateQRCode(t *testing.T) {
+	testCases := []struct {
+		name    string
+		url     string
+		wantErr bool
+	}{
+		{
+			name:    "success",
+			url:     "https://example.com",
+			wantErr: false,
+		},
+		{
+			name:    "invalid url",
+			url:     "",
+			wantErr: true,
+		},
+	}
+
+	for _, tc := range testCases {
+		service := NewService(nil, nil, 0)
+		t.Run(tc.name, func(t *testing.T) {
+			_, err := service.GenerateQRCode(tc.url)
+			if tc.wantErr {
+				assert.Error(t, err)
+			} else {
+				assert.NoError(t, err)
+			}
+
+		})
+	}
+}
